SFR-1064 Add unit tests for lambda and container handlers
This adds some tests for the new handlers added to make the conversion tool more portable.

This also includes a minor refactor of how the container API is invoked to make it testable. The core `createServer` method is not tested here but as that only invokes a single method that should not be an issue. This also adds stronger typing to the container method.