SFR-1064 New Handlers (#3)
* SFR-1064 First cut at lambda handler

* SFR-1064 Working Lambda configuration

* SFR-1064 Add Dockerfile and handler for docker process

This adds a new entrypoint for running this service in a container. In an effort to minimize the dependencies for this project the API is built with the default node `http` module.

* SFR-1064 Lint new files

* SFR-1064 Update Dockerfile and README

* don't install optionals in ci

* require module instead of importing

* move require into try block

* SFR-1064 Add unit tests for lambda and container handlers

This adds some tests for the new handlers added to make the conversion tool more portable.

This also includes a minor refactor of how the container API is invoked to make it testable. The core `createServer` method is not tested here but as that only invokes a single method that should not be an issue. This also adds stronger typing to the container method.

* SFR-1064 Linting new changes

* SFR-1064 Don't run server during tests

Co-authored-by: kristojorg <kristo.jorgenson@gmail.com>